+++ Profiler/ProfilerCommands.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using NLog;
+using Profiler.Basics;
 using Profiler.Core;
 using Profiler.Interactive;
 using Profiler.Util;
